Create Min_Jumps_to_Reach_End.java
Given an array arr[] of non-negative integers. Each array element represents the maximum length of the jumps that can be made forward from that element. This means if arr[i] = x, then we can jump any distance y such that y ≤ x. Find the minimum number of jumps to reach the end of the array starting from the first element. If an element is 0, then you cannot move through that element.

Note:  Return -1 if you can't reach the end of the array.

Solution Explanation:

Here we are first taking out the base cases, like if the arr is null or contain single element or 1st element is zero.

Then we initialize, 
maxReach to 1st element, maxReach represent, the Max Index we can reach from the current index.
remaining_Steps to 1st element, as the name suggests, as we start from 1st element, we set the remaining steps we can take from the element without adding a jump.
and then we initialize jumps to 1 with out any base cases, reaching to end will cause at least 1 jump.

Then we traverse the array and at every point we check if we reached the end or can we reach the end from the current position. 
Next we update, the Max Reaching Index from the given position. 
then we update the remaining steps, and if the remaining steps come to zero, we check if we can reach the current index from the position we started, by checking it with max reach.
Then we update the remaining steps with subtracting it from the maxReach and from current position because we need not go back to the Max Reach position, we can continue from the current index by adjusting the steps remaining.

If there are any values that comes in middle before the current jumps we update it through Max Reach.