Create Amajority_Element.java
Majority Element
Difficulty: Medium Accuracy: 27.82% Points: 4
Given an array arr. Find the majority element in the array. If no majority exists, return -1.

A majority element in an array is an element that appears strictly more than arr.size()/2 times in the array.?

Solution:
Here we are using Hash Map to count the elements, store the count in Hash Map, element as Key, count as Value pair.
And we use, getOrDefault(Key, Default) method to get the count of the number, if it is presented in the HashMap.